move EntityDefinitionComparator to common bundle
