+++ ui/plugins/eu.esdihumboldt.hale.ui.views.mapping/META-INF/MANIFEST.MF
@@ -18,6 +18,7 @@ Import-Package: com.google.common.base;version="9.0.0",
  de.fhg.igd.slf4jplus,
  eu.esdihumboldt.hale.common.align.compatibility,
  eu.esdihumboldt.hale.common.align.extension.function,
+ eu.esdihumboldt.hale.common.align.helper,
  eu.esdihumboldt.hale.common.align.model,
  eu.esdihumboldt.hale.common.align.model.impl,
  eu.esdihumboldt.hale.common.align.model.transformation,
